backend/requirements.txt
blinker==1.8.2
click==8.1.8
dnspython==2.6.1
flask==3.0.3
Flask-Cors==5.0.0
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.5
MarkupSafe==2.1.5
pymongo==4.10.1
python-dotenv==1.0.1
werkzeug==3.0.6
zipp==3.20.2
